move the URI patterns and zod schemas for tool input to the sdk
