+++ front/lib/actions/mcp_internal_actions/input_schemas.ts
@@ -1,9 +1,15 @@
-import type { InternalConfigurationMimeType } from "@dust-tt/client";
-import { assertNever, INTERNAL_MIME_TYPES } from "@dust-tt/client";
+import type {
+  ConfigurableToolInputType,
+  InternalConfigurationMimeType,
+} from "@dust-tt/client";
+import {
+  assertNever,
+  ConfigurableToolInputSchemas,
+  INTERNAL_MIME_TYPES,
+} from "@dust-tt/client";
 import { Ajv } from "ajv";
 import assert from "assert";
 import type { JSONSchema7 as JSONSchema } from "json-schema";
-import { z } from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
 
 import type { MCPToolConfigurationType } from "@app/lib/actions/mcp";
@@ -22,56 +28,6 @@ import {
   setValueAtPath,
 } from "@app/lib/utils/json_schemas";
 import type { WorkspaceType } from "@app/types";
-
-export const DATA_SOURCE_CONFIGURATION_URI_PATTERN =
-  /^data_source_configuration:\/\/dust\/w\/(\w+)\/data_source_configurations\/(\w+)$/;
-
-export const TABLE_CONFIGURATION_URI_PATTERN =
-  /^table_configuration:\/\/dust\/w\/(\w+)\/table_configurations\/(\w+)$/;
-
-// URI pattern for configuring the agent to use within an action (agent calls agent, sort of Russian doll situation).
-export const CHILD_AGENT_CONFIGURATION_URI_PATTERN =
-  /^agent:\/\/dust\/w\/(\w+)\/agents\/(\w+)$/;
-
-/**
- * Mapping between the mime types we used to identify a configurable resource and the Zod schema used to validate it.
- */
-export const ConfigurableToolInputSchemas = {
-  [INTERNAL_MIME_TYPES.CONFIGURATION.DATA_SOURCE]: z.array(
-    z.object({
-      uri: z.string().regex(DATA_SOURCE_CONFIGURATION_URI_PATTERN),
-      mimeType: z.literal(INTERNAL_MIME_TYPES.CONFIGURATION.DATA_SOURCE),
-    })
-  ),
-  [INTERNAL_MIME_TYPES.CONFIGURATION.TABLE]: z.array(
-    z.object({
-      uri: z.string().regex(TABLE_CONFIGURATION_URI_PATTERN),
-      mimeType: z.literal(INTERNAL_MIME_TYPES.CONFIGURATION.TABLE),
-    })
-  ),
-  [INTERNAL_MIME_TYPES.CONFIGURATION.CHILD_AGENT]: z.object({
-    uri: z.string().regex(CHILD_AGENT_CONFIGURATION_URI_PATTERN),
-    mimeType: z.literal(INTERNAL_MIME_TYPES.CONFIGURATION.CHILD_AGENT),
-  }),
-  [INTERNAL_MIME_TYPES.CONFIGURATION.STRING]: z.object({
-    value: z.string(),
-    mimeType: z.literal(INTERNAL_MIME_TYPES.CONFIGURATION.STRING),
-  }),
-  [INTERNAL_MIME_TYPES.CONFIGURATION.NUMBER]: z.object({
-    value: z.number(),
-    mimeType: z.literal(INTERNAL_MIME_TYPES.CONFIGURATION.NUMBER),
-  }),
-  [INTERNAL_MIME_TYPES.CONFIGURATION.BOOLEAN]: z.object({
-    value: z.boolean(),
-    mimeType: z.literal(INTERNAL_MIME_TYPES.CONFIGURATION.BOOLEAN),
-  }),
-  // We use a satisfies here to ensure that all the InternalConfigurationMimeType are covered whilst preserving the type
-  // inference in tools definitions (server.tool is templated).
-} as const satisfies Record<InternalConfigurationMimeType, z.ZodSchema>;
-
-export type ConfigurableToolInputType = z.infer<
-  (typeof ConfigurableToolInputSchemas)[InternalConfigurationMimeType]
->;
 
 /**
  * Mapping between the mime types we used to identify a configurable resource
